+++ genvalidity-hspec-binary/test/Test/Validity/BinarySpec.hs
@@ -10,6 +10,7 @@ import Test.Validity.Binary
 spec :: Spec
 spec = do
     binarySpecOnGen (genListOf $ pure 'a') "sequence of 'a's" (const [])
-    binarySpecOnValid @Double
+    -- binarySpec @Double DOES NOT HOLD
+    binarySpecOnValid @Rational
     binarySpec @Int
     binarySpecOnArbitrary @Int