+++ genvalidity/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+* `genUnchecked` of `Double` and `Float` now generates `NaN`, `+Infinity`, `-Infinity` and `-0` according to the new version of `validity`.
+
 ## [0.6.1.0] - 2018-10-06
 
 ### Changed