+++ validity/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## [Unreleased]
 
+## [0.9.0.0] - Unreleased
+
+### Added
+
+* `prettyValidate`, `validationIsValid`, `prettyValidation`
+* `validateNotNaN`, `validateNotInfinite`
+
+### Changed
+
+* Renamed `prettyValidation` to `prettyValidate` before adding the new `prettyValidation`.
+* `NaN`, `+Infinity` and `-Infinity` are now considered valid for both `Double` and `Float`.
+
 ## [0.8.0.0] - 2018-08-25
 
 ### Added